build: Partially revert #4193, the avoidance of add_blah wasn't necessary (#4273)
I misunderstood all along -- the add_compile_definitions,
add_compile_options, etc., aren't truly global, they are restricted to
targets in the same directory or below. They aren't really prone to
polluting "surrounding" projects when OIIO is used as a subproject.

So for simplicity, reverting that part and using the usual built-ins.

Signed-off-by: Larry Gritz <lg@larrygritz.com>